+++ domain/src/main/java/com/example/domain/repo/UserRepo.kt
@@ -267,7 +267,7 @@ interface UserRepo {
     )
 
     suspend fun deleteStationAlarmFromDatabase (
-        alarmID:Int,result: (Resource<String>) -> Unit
+        alarmID:Long,result: (Resource<String>) -> Unit
     )
     suspend fun updateStationAlarmFromDatabase (
         stationAlarmEntity: StationAlarmEntity,result: (Resource<String>) -> Unit
